Support MultiLines
Since an intersection with a polygon can result in potential MultiLines
in addition to Lines, we adopt an approach similar to the one used in
SummaryJob where we do all our processing on MultiLines instead. We
extract Lines from MultiLines before adding them to the STRtree.

We also add an example request with MultiLineStrings.