Fetch tiles intersecting shape, not bounding box
Previously we would fetch all tiles from S3 that intersected with
the bounding box of the incoming shape. This was wasteful, espec-
ially in cases where the shape was a long diagonal.

Now, by cropping to the multiPolygon resulting from unioning all
incoming shapes, rather than calculating an envelope, we ensure
that only relevant tiles are fetched.

Since fetching from S3 is a sizable portion of our latency, this
change improves performance 10-20%.