Allow pixels to be treated as areas on demand
Previously, we would always treat raster pixels as points,
so if a pixel's centroid was covered in an area of interest
it would be included, otherwise not. For certain rasters,
where the pixel sizes are very large, for small areas of
interest, no pixels are included since the areas do not
cover their centroids.

By allowing the calling JSON to specify whether to treat
pixels as points or areas, we allow treating certain rasters
(as determined by the caller) as having pixels as areas,
which will make small areas of interests run again.