Add RasterSummary operation
For a given shape and list of rasters, this will return a list of
{min: 0, avg: 0, max: 0} objects for each raster in order. For the
attached example, the expected output is:

    {
      "result": [
        {
          "min": -38,
          "avg": 799.9116994958924,
          "max": 3430
        },
        {
          "min": 0,
          "avg": 1.2253132788318364,
          "max": 28.67453956604004
        }
      ]
    }

In accordance with the LongAdder and DoubleAdder used for counting
and averaging in RasterGroupedAverage, we use a DoubleAccumulator
for both min and max values, with custom MinWithoutNoData and
MaxWithoutNoData implementations that ignore NoData values and
only count the rest.

This will only work with Double type layers, not with Int layers.