Mock out new multioperation endpoint
The new multi/ endpoint will take a list of shapes and
a list of operations, and will perform given operations
on the given shapes to return a single blob of results.

The actual implementation will follow, but this is to
solidify an interface, unblocking development in the
main repo.

The new endpoint can be tested with:

    http :8090/multi < examples/MultiOperationRequest.json

Refs #82