Increase request timeout to 59s
The default value is 20s, which is too short for when running a
multi-operation request inside the Worker VM. The default limit
for idle-timeout is 60s, and the value for request-timeout must
be less than that, so we set it to 59s.

We don't set it to be too long because that would tie up the
server resources. However, we do want it to be a bit longer than
what the current allowance is.

The Schuylkill HUC-8 with 61 child HUC-12s take ~45s for running
subbasin inside the Worker VM. This should give us some breathing
room for high-load situations.