Fallback to empty tile for queries out of bound
In cases when one layer has a larger extent than another, such
as NLCD and ARA, it is possible to get Spatial Keys from one
which are not available in the other. This would previously
cause a crash. Now we default to an empty tile (with NODATA
values) instead.

Also add an example request of a square shape in Phoenix, AZ
which is within NLCD but outside ARA. This request would
previously cause an Index Out of Bounds error. Now we get
back correct values, with NODATA for ARA.