Update multi to take multiple stream lines
Previously the `run` endpoint would take a list of MultiLine
strings, and the `multi` endpoint would take a single MultiLine
string. Given the way we collected the lines from the base NHD
dataset, this was fine, as there was a single MultiLine string
at the end of collection.

However, this was changed to incorporate the high resolution
streams in #3428, specifically in 4f48577. This was necessary
because the high resolution dataset is composed of MultiLines,
and collects into a GeometryCollection rather than a MultiLine.

This new style worked with the `run` endpoint, which is used for
analysis, but not with the `multi` endpoint, which is used for
modeling. To enable modeling of high resolution streams, we now
change the `multi` endpoint to also take a list of MultiLine
strings.

Also see:
  - https://github.com/WikiWatershed/model-my-watershed/pull/3428
  - https://github.com/WikiWatershed/model-my-watershed/pull/3428/commits/4f48577790b5d382ad902733ad19ce05424b536c