Add MapShed and Subbasin tests
We repurpose the benchmark scripts to save their outputs,
and then add an optional --verify parameter that will compare
the live results against these saved outputs.

Also add a test script that runs the benchmark-mapshed and
benchmark-subbasin scripts with the --verify parmeter to run
the tests.

Also add a ruff script to format the Python scripts.