Set server script output
Need to discart output of the server script in order
to fully run the process in the background.  This
may need revision as we would want to capture script
failures, but for now, testing running the server in
the background.  This discard will prevent the
process from being suspended when sent to background.